Improve flow type syntax highlighting
vim-javascript makes everything one colour (like a type), which makes it
look like some sort of comment, but I prefer to have it as a regular
language construct.
To be able to distinguish the type annotations from an object, it uses
different colours for the keys and all values use the type colour.