Rework autocommands for nvim 0.7
- autcommands can be defined with `vim.api.nvim_create_autocmd`
- augroups are easily integrated and created/cleared with
  `vim.api.nvim_create_augroup`.

Already had a similar abstraction again, just like with the mappings,
but this time there is a lot of boiler plate code that can be removed.
No awkward group creation and such is needed anymore. Just create the
autocommand and specify which group it belongs to, also creating a group
clear all autocommands.
Almost the whole utils.autocmd could be deleted, only the grouping in
a table (iterating) is really needed.