Add Hydra plugin for special modes such as resizing
Hydra originated from Emacs, which doesn't really seem to be that
helpful for Vim, but there are just a few scenarios where this is
exactly what I need for things that were just too annoying, such as
resizing windows.
This allows to have a `resize mode`, where when activated it can be
continuously resized by hitting `hjkl` instead of having extra bindings.
Also the implementation is very clever, seems so elegant as well.
I tend to use `pink` heads, which allow other keys that are not defined
within the mode to be passed through, which allows things like moving
across windows while remaining in resize mode, meaning that it does not
need to be called again in the next window. The downside to this would
be that you have to cancel it by hitting `q` or `Esc` instead of it
being automatically disabled once you switch windows for example.
Overall this is my preferred way as it feels the most natural to my
usage patterns.
It's just one of these plugins that overcome one specific annoyance.