Modernise NeoVim look with global statusline and noice
Noice brings some modern elements that are present in many other editors
such as command palette and notifications, which kind of makes a global
status line more appropriate, which is supported by NeoVim with
`set laststatus=3` and lualine also supports it. Also winbar can be
added but because it uses up an entire row, a floating window plugin is
used instead to show the buffer name at the top of each window.

Main changes:
  - Global statusline
  - Virtual winbar (not actual winbar) for buffer names
  - Noice: Command palette, notifications, Lsp progress etc.
  - Updated Theme as the high contrast statusline doesn't make sense
    with the new look.

Note: All the floating windows that are being created (and destroyed
pretty much right afterwards) increases the buffer number heavily, which
doesn't play well with the buffer numbers in the bufferline and was one
criticism of Telescope, so it might be worth looking at it again,
especially since the notifications kind of rely on Telescope to inspect
them properly.