Include which-key plugin
I mainly need it for the spell checking `z=`, as Noice is quite terrible
with it and which-key is the proposed solution.
It's also nice for registers and marks, to show a preview.
I'll have to see how I like the other suggestions, as they are on
a timeout I think I'll rarely see them, and they might even be nice from
time to time, otherwise I'll disable them later on.