Use nvim-surround instead of vim-sandwich
I still think sandwich has some more features, but as I'm using these
from very rarely to never (such as double count mechanism), the newer
nvim-surround is just more appropriate, especially as it fixes one bug,
where the undo had to be done in two steps instead of one, and the
option of not moving the cursor after the action was performed is just
a big plus, as that's most noticeable as it's most used.

I'm staying with the sandwich mappings, as I've got used to them and
find them more logical after a reluctant switch from the original
vim-surround.