Switch to Telescope from Fzf
Telescope is now usable, especially thanks to the fzf-native extension,
which makes the matching work (finally) as expected, as it uses the same
algorithm as fzf (it's a C port of Fzf).
The biggest annoyances that remains is the sheer excess of buffers being
created, but as I've already switched to plugins that create a lot of
floating windows, I'll move away from using buffer numbers.