Only show buffers that were used within the tab page
I was used to that previously, and I just prefer it, as I never use
a tab page as a different view, but rather as a separate environment.
Because that uses a plugin that scopes them, they are also not
searchable with a fuzzy finder, which is different from the previous
version I was used to, but it still is better overall.