Expose nvim theme as colorscheme mine
Got annoyed that Lazy sets a dark theme (habamax) when it is popping up
directly at the start, then I either needed to reopen nvim or call the
lua function manually, much easier with `colorscheme mine`.
Also it is possible to specify the colorscheme that Lazy uses, so for
that it needs to be exposed as such.