Make nvim split resizing behave more nautrally (like tmux)
It's essentially the last buffer that is unnatural by default, because
making it bigger expands it to the left/up, but making it bigger is
usually mapped to the other side (l / k) as that is the direction the
windows grow, except for the last ones.
Now it behaves to always resize the current window in the direction,
regardless of whether it makes the window bigger or smaller.