feat(nvim): add a scrollbar with nvim-scrollview
Supports some nice signs next to it, making it very useful, and it looks
good in general and fits in quite nicely.
Only the focused window shows the scrollbar, so that makes it also
easier to identify which window is focused, which got a bit more
difficult since switching to global status line.