fix(nvim): replace right bar with SF Mono compatible
I've used *Right One Quarter Block* which seems to be part of a legacy
unicode character block, which does not exist in SF Mono.
So had to move to *Right Half Block* because *Right One Eighth Block* is
too small to see. The 1/4 would be ideal, but unfortunately, it no
longer exists, even though the left version does.