feat(nvim): include flash for jumping around
It's like Leap/Hop etc. but is more to my liking. Just having the labels
for the search would have been great, but it has a major caveat, reason:

Don't want to have labels on the regular search, while it is quite cool and makes it especially easy to use
(not new keybind, just an additional feature during the search), there is one crucial drawback, namely that
it will automatically jump (and cancel the search afterwards) if something is typed where it hits a label
(i.e. when the search term doesn't exist, but part of it did, so the label selected that instead of
continuing to narrow the search).
I would even prefer this to the regular jump (can be used for operator pending as well), just because you can
see what you typed for the search. But that should only activate from a mapping.

But also the Treesitter selection is what I was looking into and that
was especially nice, compared to the built-in incremental selection of
treesitter. It happened to be made by f0lke just 2 days earlier.