feat(nvim): create some useful Python snippets
I used to have something like `cls` but that doesn't exist with the
snippets plugins I use now.
Also the one that I used a lot was `dbg` to have my custom debugger.
And now also a dataclass, because I realised that I use them quite a lot
but creating them was always annoying.