fix(nvim): configure nvim-lightbulb to use the same icon
It used to be that you need to use `vim.fn.sign_define` but that had
been moved to the setup call instead. This was the case for a while, but
since I've used the same icon it wasn't an issue, also it being that
specific one was okay since it made it distinct to the Hint diagnostic.
But since that now causes some alignment issues with some fonts,
I needed to change it.