feat(nvim): replace null-ls with conform
Conform is for *formatting* where the LSP does not have a formatter or
not the correct one, which was the main purpose of having null-ls.
It is a simple as `calling.format` and if a formatter is defined in the
options of conform, it will run that just like the LSP formatter would
(partial edits etc.) or fallback to the actual language server.

It is much nicer to integrate without having to manually disable the
language server formatter on attach, because you would have to choose
between null-ls and the LSP.
Much more straight forward and null-ls is no longer maintained so it
needed to be replaced at some point.