feat(nvim): add glance to preview LSP definitions/references etc.
It's really nice to look up the definitions and references, including of
their type (as long as it's supported by LSP, e.g. Rust does, but is one
of the rarer ones).
The window is very practical and you can even edit the code directly
from there if necessary.
All in all, very nice plugin, even though I might not end up using it
all that often.