feat(nvim): enable teesitter text objects
Particularly functions/class objects for selection and navigation, but
also to swap parameters.
Swapping parameters with `[a` and `]a` respectively, conflicted with
existing mappings for the arglist navigation, but since I never use an
arglist, I just removed these mappings. Not entirely happy with the
bracket mappings for this, as they are usually to jump, but nothing else
feels good either, so for now it will be like this.