feat(nvim): format Python with ruff's new formatter
Ruff has now a formatter, so black can be replaced by it.
Unfortunately, the isort replacement (sorting imports) that was already
present is not part of the formatter, but the linter with `--fix`,
meaning that the user needs to run both (fix and format) to format them.
But that at least makes it possible to remove isort and black all
together and only use Ruff.
Hopefully, in future updates the formatting will be unified, but since
this is the very first release (it technically is even a beta), there
are some rough edges. I still prefer to adopt it immediately.

In order to run both the fixer (for import sorting) and the formatter,
I decided to do it with conform. Since the import sorting lint is not
enable by default, the formatting wouldn't kick in, so therefore
I enforced the lint rule when running the fixer in conform.

The additional benefit of running it through conform instead of the
ruff-lsp, is that the formatting only adds the edits rather than fully
replacing the whole file every time. This preserves things like marks
and also the undo history is much nicer.