feat(nvim): enable inlay hints (plugin before v0.10)
The inlay hints are enabled through the plugin, which displays them at
the end of the line, rather than inline.
It will be replaced with the native implementation that is coming to
nvim v0.10.
I used to have inlay hints for Rust through rust-tools, but they have
been removed in rusteacenvim and they refer to that plugin.