feat(nvim): jump outside of treesitter node in insert mode
Most useful to jump outside of curly brackets, especially multiline
ones.
Very much inspired by smart-tab.nvim or tabout.nvim, but more aligned to
my taste, where <C-h> jumps to the beginning of the node (outside of it)
and <C-l> to the end. It does it somewhat smart, i.e. when the cursor is
at the beginning of wrapped node (inner node) it jumps to the actual
parent expression (not just the wrapper).
Also moved the snippet forward into a regular mapping instead of in the
cmp definition (is used as the fallback).