nvim(feat): replace exchange with subtitute plugin
It does the same as exchange, but also adds a feature I didn't know
I wanted, namely that you can paste to overwrite text from a motion.
The use case I encounter sometimes, is that I want to replace the
content of a string i.e. in quotes such as "some string" with the yanked
text. For this I would usually select it then paste.
With this it is just `sfi"` (`sf` is the binding) and it replaces it
without overwriting the yank register.