chore(nvim): set default type checking mode of basedpyright
The default is very strict and it is pretty much unusable on either an
untyped code base or just a lower quality one in general, because there
are a thousand errors.
So set it to standard by default, which makes it usable on these code
bases, and good code bases have them configured anyway.