feat(nvim): use real inlay hints of nvim 0.10.0
Neovim 0.10.0 now has support for inlay hints, where virtual text is
shown inline.
That means the plugin is now obsolte, and it has already been archived.
Even though I'd still like the option to show the virtual text at the
end, as the plugin did. But they don't play well together and I'll just
get used to it for now, and hope that down the line that feature will be
added as an option.