feat(nvim): switch to Neogit (+diffview) for git integration
Finally replacing fugitive. While fugitive is still one of my favourite
plugins I've ever used, Neogit just seems more modern, and it solves
some of the minor annoyances from fugitive I had, which are mostly when
pre-commit takes a while, there was no feedback (here there is a popup)
and sometimes the git repo was not correct when editing config files.

I almost gave up on Neogit, because it was an absolute nightmare to
configure, as it has some very stupid design decision, such as mappings
causing all sorts of trouble (overwriting navigation keys...), and all
highlights being super annoying and having to redefine like 20 groups.

But I pretty much found workarounds for everything I care about, and the
experience is very similar to fugitive, which makes my general workflow
the same, but with the new neogit features.