fix(nvim): link highlight redo and undo to same highlight
It was added to distinguish between undo and redo, but I don't have
another colour, and also I don't see the point in having a different
highlight, since I know what I pressed.
If it would be what is added / deleted when doing an undo, that would be
a different story.