feat(nvim): replace telescope with fzf-lua
I always had issues with telescope and lately it is really the speed in
repos that have a lot of artefacts, even if they are not tracked by git,
even a git files got really slow. I always preferred fzf, but that was
missing some features. But now fzf-lua provides all of it and is much
closer to what I like.
Some of the configuration was a bit annoying, but it just seems to do
the small things nicely that makes using it more enjoyable.