feat(nvim): show inlay hints at end of line with lsp-endhints
The inlay hints with nvim 0.10 are nice when reading code, as they are
displayed inline such that it looks as if a type definitions were
present even if it's just inferred. But when writing code, it's really
annoying since it makes the cursor jump a lot, which I hate. So I just
always disabled them during insert mode, but that sometimes bugged them
out. Also I'd like to see them hints in insert mode, but without being
distracted.
Displaying them at the end with the plugin lsp-endhints is in most cases
the better solution. And this plugin compared to the old (now archived)
one, allows to switch between the native inline nvim inlay hints and the
end of line hints, therefore getting both of with a simple toggling
mapping.