feat(nvim): replace render-markdown with markview
Seems a bit more modern and more active as well. The split mode is
really something I like.
On the other hand, the unconcealing of the title is a bit jumpy, I don't
like that, but I think overall it has more potential, let's hope that
a future update fixes this annoyance.