fix(nvim): switch to a perf fork for cmp
I thought going to magazine was temporarily the best option, as this was
built on top of the perf fork, while also merging additional PRs.
But within a day of use, I found a bug that makes it impossible to use,
so I had to change to just the performance branch, since it makes it
unusable. I cannot go down in the completion menu with <C-j>, which just
inserts a newline.
Unbelievable, I never encountered any bugs with cmp, and this just broek
immediately.