feat(nvim): use snacks.nvim including notification instead of nvim-notify
I mainly wanted it for the bigfile handling, as that was something
I have been seeing lately, which I thought that I surely need to add
one, just to cope with opening large JSON files.

But then I also used its notifications instead of nvim-notify, as I had
no strong attachment to it and it seems to do the same job, also since
it's become the default in noice, that seems like a no-brainer.