feat(nvim): refresh neogit status buffer on focus/enter
Again baffling how that is not the default. Neogit just only refreshes
the status buffer if an action has been performed through certain
mappings.
I've added two cases to refresh automatically;

- When neovim is focued (i.e. other shell has changed something)
- When entering the status buffer

There are other cases that are not automatically picked up, but at least
when entering it, it will always be up to date.
How is this not integrated in Neogit?