fix(nvim): restore python module import highlights
It seems that a new version of the Python Tree Sitter parser highlights
the module names in the import statements. I don't like it because it
blends in with the keywords, and even if so, it is completely pointless
since in that exact case it can never be anything else.