feat(nvim): replace nvim-cmp with blink.cmp
Much faster, more modern. Will still need to tweak the configs a bit,
but so far I at least got the look that I wanted and it is indeed fast.
As well as having a better ordering, as sometimes nvim-cmp got annoying
to show me a lot of stuff that I don't care about, such as the whole
double underscore methods in Python.