feat(nvim): create dedicatd mapping to jump start/end of tresitter node
I had it integrated with the snippet jumps, but occasionally I triggered
it accidentally because I though I still had a jump left. So a dedicated
mapping is a better idea, and I noticed the Mac (or emacs) binds for
start/end of line were sort of free in insert mode (they have something
completely useless by default).