feat(git): enable autoSqaush and updateRefs for rebase
Very useful to rebase fixups into previous commits, including when other
branches build on top of the branch that is being rebased with the
fixes.
I don't see any scenario where I would want to have them disabled, so
that's why they have been added to the git config.

They can always be turned of with the options `--no-autosquash` and
`--no-update-refs`, respectively.