chore(nvim): upgrade plugins to always show ghost_text in blink.cmp
This allows to always show the ghost_text without having the menu open.
I quite like it this way, but still need to decide what I want exactly.
Also currently has a bug of not being able to open the menu, but for now
it is a local workaround waiting for it to get resolved.