fix(Makefile): include hyphened targets and align the help message correctly
For the help message, there was a tab added between the name of the
target and the description, and afterwards `columns` aligns it from
after the tab. That means that if it is longer than one tab stop, it
will be offset by one additional tab. So let `columns` handle this.