feat(nvim): enable image preview with snacks.nvim
The support was added a little bit ago, and unlike image.nvim, there is
no dependency on lua rocks, which makes it much more usable.
There is no way to toggle the rendering, unfortunately, which was the
main reason I never added image.nvim permanently.
But I'll just get used to it.