fix(nvim): enable treesitter correctly for languages with multiple filetypes
For example, the latex treesitter parser was not enabled for tex files,
even though the same one is used for both. So just needed a mapping to
all of them, which was just simplest to create an autocmd per filetype
rather than a global one, otherwise it would be a lot of unnecessary
lookups during the autocmd.