chore(nvim): disable tab to jump in snippets
I have `<C-h>` and `<C-l>` to jump and Tab should just insert a tab, but
for some reason, not only does blink automatically add this, but also
the default nvim bindings now have it.

I noticed this when I was doing some Typst, because for some reason
I was longer in a snippet than I would have liked (due to the LSP) and
then I just couldn't insert a tab.